feat(dashboard-ui): expose the findings view vocabulary as its own subpath
`./findings/views`, for the same reason `./theme` and `./recommendations` exist:
a host that needs only the view vocabulary should not pay for the barrel's ~80
exports, which include Radix-backed dialogs and d3-shape charts.

The case is not hypothetical. A route's `validateSearch` runs at route-tree
construction, so any module it reaches is eager by nature — and reaching
DEFAULT_FINDINGS_VIEW and isFindingsView through the barrel put the whole chart
surface on the critical path of a page that renders none of it.

The vocabulary moves into views.ts, which imports NOTHING, and that is the whole
property rather than tidiness: a subpath is only worth the name if it is a leaf.
Pointing one at meta.ts would still have dragged a ui-kit tone and the icon set
behind it — measured at 725 kB of eager JavaScript against 548 kB for the leaf.
meta.ts re-exports all four symbols, so every existing import is unaffected.
